refactor: route component imports through facades instead of direct services
