+++ src/components/settings-update.js
@@ -5,7 +5,7 @@
 import { _el } from '../utils/dom.js';
 import { createSettingsSection } from '../utils/settings-section-builder.js';
 import { registerComponent } from '../utils/component-registry.js';
-import updateApi from '../services/update-api.js';
+import { updateFacade as updateApi } from '../facades/update-facade.js';
 
 function _showCheckButton(area, onCheck) {
   area.replaceChildren();
